model/assessment/requirements.txt
scikit-learn
pandas
numpy
pyyaml